pyproject.toml
[tool.poetry]
# Change this line
name = "current-weather-cli"
version = "0.1.0"
description = "A simple CLI to get the current weather from an API."
authors = ["awakra <lucasadls0@gmail.com>"]
readme = "README.md"
packages = [{include = "current_weather_cli"}]

[tool.poetry.dependencies]
python = ">=3.12"
requests = ">=2.32.4,<3.0.0"
typer = "^0.16.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"

[tool.poetry.scripts]
weather = "current_weather_cli.main:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"